chore(model): re-organize pyannote.audio.models
