+++ pyannote/audio/cli/train_config/model/PyanNet.yaml
@@ -1,9 +1,2 @@
 # @package _group_
 _target_: pyannote.audio.models.segmentation.PyanNet
-lstm:
-  hidden_size: 128
-  num_layers: 2
-  bidirectional: True
-linear:
-  hidden_size: 128
-  num_layers: 2
