+++ pyannote/audio/cli/train_config/model/debug.yaml
@@ -1,2 +1,2 @@
 # @package _group_
-_target_: pyannote.audio.models.debug.SimpleSegmentationModel
+_target_: pyannote.audio.models.segmentation.debug.SimpleSegmentationModel
