+++ pyannote/audio/models/embedding/__init__.py
